♻️ Move collection files into folder
